refactor: extract types, client, and interactive logic for multi-template support
Phase 1 Foundation refactoring for issue #15:
- Extract types to internal/templates/types.go
- Extract API client to internal/templates/client.go with FetchTemplates/RenderTemplate
- Create cmd/render_types.go with RenderConfig, GitConfig, PRConfig, RenderResult
- Move interactive form logic to cmd/render_interactive.go
- Slim down cmd/render.go to entry point with flags and routing
- Add unit tests for templates client

Closes #15
